pyproject.toml
[project]
name = "shmoopbot"
version = "0.2.0"
description = "A Discord bot for the Fire Totem Games Discord server"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "discord-py>=2.7.1",
    "python-dotenv>=1.0.1",
]
